nova-platform/requirements.txt
numpy>=1.23
PyYAML>=6.0
protobuf>=4.21
networkx>=2.8
tqdm>=4.64
